+++ packages/plugin-social-polls/web/src/createPollPlugin.ts
@@ -15,9 +15,8 @@ const createPollPlugin: CreatePluginFunction<PollPluginEditorConfig> = config =>
     theme,
     type: POLL_TYPE,
     toolbar: createToolbar({
-      helpers,
       isMobile,
-      theme,
+      settings,
       t,
     }),
     helpers,