+++ packages/plugin-social-polls/web/src/toolbar/createToolbar.ts
@@ -70,13 +70,15 @@ const externalToolbarStyles: ModalStyles = {
   },
 };
 
-export const createToolbar: CreatePluginToolbar = ({ isMobile, t }) => {
+export const createToolbar: CreatePluginToolbar = ({ isMobile, settings, t }) => {
+  const { enableVoteRole } = settings || {};
+  const config = { ...DEFAULT_COMPONENT_DATA.config, enableVoteRole };
   const buttonProps = {
     type: BUTTON_TYPES.MODAL,
     name: INSERT_PLUGIN_BUTTONS.POLLS,
     tooltip: t('Poll_InsertPoll_Tooltip'),
     getIcon: () => InsertPluginIcon,
-    componentData: DEFAULT_COMPONENT_DATA,
+    componentData: { ...DEFAULT_COMPONENT_DATA, config },
     modalElement: decorateComponentWithProps(PollPresetSelector),
   };